Extract `FeedLoaderStub` to a shared scope
