+++ FeedFetchApp/FeedFetchAppTests/FeedLoaderWithFallbackCompositeTests.swift
@@ -72,18 +72,3 @@ private extension FeedLoaderWithFallbackCompositeTests{
     }        
     
 }
-
-// MARK:- FeedLoaderStub
-
-private final class FeedLoaderStub: FeedLoader {
-    private let result: FeedLoader.Result
-    
-    init(result: FeedLoader.Result) {
-        self.result = result
-    }
-        
-    func load(completion: @escaping (FeedLoader.Result) -> Void) {
-        completion(result)
-    }
-    
-}
